separating app component into its own file, displaying actual image from url
